News Page
Added a news page to replace the placeholder. Push notifications will now be displayed on the news page.

Added storage related classes to store notifications when user recieves a push notification. Storage is achieved through local storage on disk as seen with the updated dependencies.

Updated the number on the text contact button to match the format of IOS and Android phone numbers.